Change bstr to a bstr_t rather than a void*.
The central issue here is that bstr's are *always* passed as 'bstr *'.  With a bstr as a bstr* this means passing void**'s around which are then case (via bstr_ptr) to a (bstr_t*).  Casting a double pointer to a single pointer is a recipe for confusion, both to humans and compilers.

Arguably, we should be passing bstr's around as void*s (to avoid users directly using the implementation) but this would require casting bstr to 'void' which is also a recipe for confusion.  A potentially good solution would be to change all 'bstr *'s to 'bstr' in the code and then typedef bstr as void*.  However, this would require wide spread change in the code.