IronBee++/Module: Expand Module to allow for singular Modules and testing singularity.
A singular Module is one that does not refer to anything and can be created with the default constructor.  This allows Modules to be stored in, e.g., standard containers.

Testing singularity can be done by evaluating a Module in a boolean context, e.g., if (module) ....  This is a tricky area of C++ due to bool's implicitly converting to ints.  The code uses a standard trick to avoid such incoherence.