cfgmap: Remove default values from config map init entries (see below).
When a parentless (top level) context was created the gcdata (global configuration data) of the module is copied in to the module-context configuration data.  Previously, this data was then overwritten by any cfgmap_init entries using the default values in those entries.  Thus, any gcdata that corresponded to cfgmap data was effectively ignored.  Besides adding complexity, this could be confusing to module writers as the use of gcdata for defaults depended silently on cfgmap entries.

This change removes default values for cfgmap_init entries.  Now the gcdata is used for defaults in all cases.  This simplifies the code, avoids some unnecessary copying, and removes a gotcha for module writers.

All existing modules are updated to use the new format.