Refactoring engine.h partially into engine_types.h and parsed_content.h.
This change is two-fold. First, engine.h is refactored into
engine_types.h. This allows for cleaner definitions using ib_tx_t in
parsed_content.h.

The changes in parsed_content.h and state_notify.h are to the list
representation of headers. A header list is now a linked list that is
built using a temporary wrapper object. Iteration is along a linked
list, allowing for a more typical iterator-based traversal, but building
the list is still O(1) per element added.