RNS-427 Checking in the structural changes so unblock C++ work.
This change is the header/structure portion of a larger change. That is,
the hook dispatch functions (ib_state_notify_request_headers_data, for
instance) will populate tx->request_headers and tx->request_line so that
when the context is determined the header and request information can be
presented to the appropriate callbacks.