The block action now sets flags for the engine and other rules to consume.
When an advisory block is triggered, TX.BLOCK is set to 1. Also
the tx.flags value has the IB_TX_BLOCK_ADVISE flag set.

When an immediate block is triggered the tx.flags IB_TX_BLOCK_IMMEDIATE
flag is set.

When a phase block is triggered the tx.flags IB_TX_BLOCK_PHASE flag is
set.