Adding filtering of list types by pcre to DPI.
This also includes a minor fix to the core_data_get routine to
respect the length of the string when searching for the ':'.

Longer term, the logic in core_data_get will be moved into data.c's
implementation of variable expansion so that the plugin contains less
logic and the core treatment of that DPI contains more.