Protecting against PCRE JIT-compiled code running w/o PCRE JIT available.
It is possible for IronBee's PCRE module to be compiled with PCRE JIT
but dynamically linked against a PCRE library that does not support PCRE
JIT. This change protects against that situation.

This change also includes an optimization of multithreading. We
previously copied the pattern and study data. This is not necessary as
the patter and study data are read-only. For threading PCRE JIT requires
an independent execution stack, but a copy of the machine definitions.