Putting in place new typing system for fields.
In some ways this change is a large refactoring of an implicit type
system in which strings are converted to integer. With the introduction
of a floating point type (ib_float_t) the conversion code needs to be
centralized, which is the thrust of this change.

The updated operators are currently only eq and ne. Future checkins
will update all numeric operators to use the new typing calls.

Code in core_operators.c is expected to be significantly reduced.
The extra type handling will slightly reduce performance caused by type
conversion code. To avoid this performance penalty, avoid type
conversions.

Finally, ib_ftype_unum is slated for removal. It will be removed
as part of this series of updated.