core.c: Improve coherency of ib_auditlog_add_part_header (see below).
The function allows for log->tx to be NULL in one location but then requires it to be non-NULL in many others.  This change removes the if-null case and asserts that log->tx is not NULL.

If it turns out that log->tx might be NULL, the whole function needs to be heavily reworked.