htp_utf8_decoder.c: Remove 'inline' keyword for two functions that have external linkage but use statics.
An inline function must have access to any symbols its definition uses.  In this case, it used symbols that did not have external linkage.  But the functions did have external linkage.  So any usage of them outside this file could not possibly be inlined as they would not have access to the symbols they needed.