various: Overhaul operators (see below).
Operators now have no concept of rule engine.  Some of what was in the operator code, such as inversion, is now part of the rule engine.  Operators are given a tx instead of a rule_exec and use a void* for operator specific instance data.  ib_operator_inst_t is now ib_rule_operator_inst_t.