Use fhold when returning with an unparsed character from sub-machines.
Ragel, at the end of input, checks for p == pe. If a sub-machine fails
to parse a character at position p and does NOT use fhold, the following
fret to return to the calling machine will also advance p to p+1. If p
is at pe, that is, p is the last character in the input, Ragel will set
p equal to p+1 which is past the pe pointer. Thus, Ragel will never find
the end of the input because p will never be equal to pe.