Bug fix of configuration parser causing config files longer than 8k to lose a line or exhaust the 1k token-building buffer.
The observed defect was that when parsing a configuration file longer
than 8k parsing would stop at the 8k boundary. This boundary happend to
land in the middle of a quoted string. When the next chunk was
presented Ragel the ib_cfgparser_t.fsm structure was incorrectly
reinitialized (by code generated by %% write init in config-parser.rl).

Parsing then resumed, not in the sub-machine parsing the quoted string,
but in the main Ragel machine. When the closing quote was observed by
Ragel it entered the quoted string parsing machine and began to build an
enormous string which eventually exceeded 1k and resulted in an error.

There are some other fixes, but the parser state is the critical one.