Adding optimization to vector library to prevent buffer shrinking.
Vector resizes its databuffer if it is under 1/4 full or going to
overflow. There may be uses where the user builds up a large buffer
and truncates it only to rebuild it again. In such cases it is wasteful
to release the memory back to the parent memory pool only to request it
soon after. Adding the flag IB_VECTOR_NEVER_SHRINK now prevents
shrinking the size of the databuffer.