Adding ib_resource_pool_flush().
This function is intended for use in situations where all existing
resources are invalidated. For example, when IronBee's configuration
changes any resources using a previous configuration must be able to be
discarded and replaced.

It is the user's responsibility to ensure that all resources are
returned to the pool before calling flush so that all resources are
properly flushed.