core.c: Moving the writing of auditing information later in the event sequence.
This fixes a defect in which the rule engine would write that it had
written an audit log to a shared structure _after_ core's auditing code
had read the shared value and found no auditable information reported.
Auditing is now done in the postprocessing phase.

Adding unit test that will eventually check for logging audit logs.