clipp/hash_to_pb: Fix incompatibility with Ruby 2.0 (see below).
Script was using String#size to calculate size *in bytes*.  This was fine in 1.8 and 1.9, but in 2.0 strings become unicode by default so if the value had a valid unicode in it the String#size could return a value less than the size in bytes.  This was fixed by suing String#bytesize which works in all versions.