build: Prioritize Lua headers in libs directory over any system Lua headers. (See below)
Before this commit there was a problem if Lua was installed in a non-standard location.  In that case, the build would add -I flags to AM_CPPFLAGS to find those non-standard header files (e.g., for boost) but this would cause the Lua header files there to be used instead of the ones in the source tree.  This commit rearranges Makefiles to move the local -I before any added by the configure script.