modlua: Lua modules (which are managed by modlua) now may declare and store configuration data in C using Lua JIT's LuaFFI.
This is done by declaring (ffi.cdef) a custom struct for each module and
storing it in the module index space like a normal module. Each Lua
module is allocated an ib_module_t struct which is registered with the
ib_engine_t.

This change exposed some imprecision about which module (the ib_module_t
for ibmod_lua or the ib_module_t for the LuaModuleLoad'ed file). These
defects are fixed and are a major reason for the size of this diff.