engine/rule_engine: Add rule instrumentation code (see below). (RNS-283)
Only enabled if IB_RULE_TRACE is defined; easiest way to do that is configure with —enable-rule-trace.

Tracks, per transaction, the number of times a rule fires and the amount of time spent evaluating/firing it.  Can be turned on/off on a per-rule basis.

At present, there is no user way to enable or access this information for users.  That’s coming soon as additional directives in the rules module.