modlua: Internal API clean up and fixing a memory leak with Lua module configuration directives.
The memory leak is a use of malloc with no free that has been in our
code for some time. It was recently highlighted by static analysis.

The API clean relates to a boolean "config_time". This magic value has
been renamed to "register_directives" and is now only used for calls to
static functions in lua.c. The renaming more clearly reflects what the
boolean does, which is push a function pointer to register directives or
push NULL as a placeholder.