operator: Split operators into non-stream and stream operators.  (See below) RNS-792
This commit just splits the namespaces.  Stream and non-stream operators are otherwise identical, i.e., same function type.  The separate namespaces replaces with operator capability flags for stream and non-stream.

The Lua code is broken.  Sam will fix in the development branch.

This also disables many “stream” operators in moddevel as that code is being rewritten.