autoconf/build: Fix broken configuring of ossp-uuid.
Previously on not specifying a path for uuid-config, ironbee would search
a fixed set of locations for uuid-config, rather than checking $PATH.  On
specifying a path using --with-ossp-uuid, it would trial by trying to
locate it within "sbin", "bin" under the directory specified by the user.

This has been fixed now.

--with-ossp-uuid has now been changed to --with-uuid-config, which now
expects a path to the directory containing the uuid-config binary.

Likewise, not specifying a path using --with-uuid-config, would make
configure search the $PATH variable.