autoconf/build: Fix yajl configure.
Ironbee seems to have a dependency on yajl, but configure doesn't
warn/error out if it fails to locate yajl on the system.
Also ironbee would build only if yajl is located in the default header/lib
search path.

This is now fixed.

--with-yajl has now been removed and we now have 2 options,
1. --with-yajl-includes=PATH
2. --with-yajl-libraries=PATH

Also not specifying any of these paths, would make autoconf use the
default paths.