var: Allow access to the ib_var_source_t in an ib_var_target_t.
This allows user code to better understand what value the target may be
fetching (a list or a scaler, for instance) as ib_var_target_get only
returns lists of things.

This also allows parsing of a target string to remain in var.c and not
be reimplemented by the caller, who would then directly acquire and
query an ib_var_source_t.