xrules: Adding event creation to xrule action matches.
To support this change, the directive XRuleGenerateEvent (on|off) is added
which can enable event generation. By default, event generation is off.
Event generation is disabled during the handle_response phase to avoid
late events being generated when other modules may be trying to
aggregate events.

XRuleEventTag checks now ignore any tag that starts with "xrule/" to
prevent event creation loops. If the user constructs an XRuleEventTag
rule with an argument of "xrule", it will be created but the argument
"xrule" will never match an event tag.

Part of this change exposed a bug that XRuleEventTags were being
evaluated during the logging phase, not during the logevent event.
This is fixed.