Fix double Content-Length issue.
I've seen situtations where a server will send duplicate Content-Length
headers. I've tested this with Chrome and it will ignore the second header
if the value is the same. If the value is different Chrome throws an error.

This causes a discrepency between how a browser will handle the content
and how libhtp does. To address this I've added support for checking for
duplicate Content-Length headers and ignoring subsequent ones if the value
is the same as the first one.

Also, fix up the test (74-response-multiple-cl) now that this is not an
error condition. Also, add a test where there is a value mismatch which
is a failure.