ee_oper.c: On partial matches of the ee_match operator, re-start searching until a full match is found. RNS-1211.
The previous behavior would return no-match if the first match was a
partial match. This prevented the string "ab" matching the pattern "ab"
if the match patterns includes also the pattern "a".

In this case the string "a" will be matched but it is a partial match.
This change restarts the match to allow finding the match "ab" which is
then determined to be a full match.