utf8.cpp: Normalization and UTF-8 conversion to ASCII both will now emit error characters when invalid UTF-8 is encountered, RNS-1368.
When normalizeUtf8 is called and encounteres a series of invalid UTF-8
bytes, it will emit the UTF-8 replacement character (U+FFFE) for the
first byte in the seqeunce and resume normalizing at the second byte in
the sequence (or simply the next byte in the stream if the sequence is
only 1 character long).

When converting UTF-8 to ASCII the character 0x00 is emitted.