standard_ironbee.cpp: Fixed bug where rule_id was being used as the rule_version in event generation. RNS-1519.
Also, if integers are used for GenEvent's action or type arguments,
they will be converted into their enum values. This is to support cases
where Lua is generating the predicate expression. In such cases it is
less stable to use the sybmolic string for an enum, such as
"OBSERVATION" when the enum values are immediately available.