ts_event.c: On TS_EVENT_HTTP_SEND_REQUEST_HDR, if we are blocking a request, make a best-effort to zero-out any request body that may be sent to the server. RNS-1530.
In this change, the headers are sent, but the body is zero'ed and the
Content-Length header is set to zero. It is possible that exceeding the
amount of request body that can be buffered may render this insufficient
protection in all cases.

This is a manual pull-down from the 0.12.x branch rev 370f7d61e838e5de2b8b1a8907992814c8af304d.