ts_event.c: Protect origin server from attacks in the body of a POST.
This change hooks TS_EVENT_HTTP_SEND_REQUEST_HDR to error-out of a
request before the body is sent to the origin server. The
SEND_REQUEST_HDR handling code tells IronBee to inspect and finish the
body request if it has not already done so and the request is not
already blocked.

If the request becomes blocked by this inspection attempt, then
then TS_EVENT_SEND_RESPONSE_HDR is hooked which will send the client a
proper block page. The HTTP transaction is then reenabled with
TS_EVENT_HTTP_ERROR to stop the request from going to the origin server
and to send the user the error page.