clippscript.rb: Allow for multiple headers with the same name.
If you pass a headers hash to a request or a response that has an
array as the header value, each value will be inserted as a separate
header with the same name.

For example:

  transaction do |t|
    t.request(
      ...
      headers: { 'Hdr' => [ 'Val1', 'Val2' ] }
      ...
    )
  end

Will create a transaction with two "Hdr" headers,
the first with "Val1" and the second with "Val2".