standard_ironbee.cpp: SetPredicateVar: When child1 is finished but the result has not yet been captured, capture it before setting the vars and evaluating child2.
There was a bug in which if child1 was a literal it was never
not-finished. SetPredicateVar relied on child1 being unfinished and
transitioning to finished to capture the value of child1. This edge
detection scheme fails when child1 is always finished (never
unfinished).

This change introduces a boolean to the captured values recorded from
child1 which notes the captured values as uninitialized. The, if child1
is finished the captured fields are checked to see if they are
initialized yet or not. If they are not, the finished values from child1
are captured.

Another method would be to always fetch values from child1 if it is
finished and never capture them. This method would leak memory creating
temporary holding structures over the course of a transaction and so
values are captured once.