eval.cpp: Do not return the NodeEvalState& after a call to GraphEvalState::eval(). RNS-1654.
By removing the return value from GraphEvalState::eval() we force the
caller to fetch the NodeEvalState& themselves, which reevaluates the
forwarding chain which GraphEvalState::eval() may have altered.

Commit ed673b17815c726aa67be7e4d37ca0b06f731e2e introduced lazy
NodeEvalState initialization. This also included what was believed to be
an innocuous optimization of returning the NodeEvalState after every
call to GraphEvalState::eval().

This optimization introduced a bug where a node N would be evaluated
when it was not forwarding, then it would begin to forward to another
node, but the original NodeEvalState was returned for N. The calling
code of GraphEvalState::eval() at that point would have the wrong
NodeEvalState. This caused bug RNS-1654 in which a rule with an (if...)
predicate as its root never returned a value to the rule engine and so
was never non-null and never fired.